Very rudimentary particle system support
Also highly wasteful at the moment since it's allocating huge piles of
memory statically. I'll trim it down and have it dynamically allocate
once I've got the functionality working.

So far there are three types of particle system: points, lines, and
lightning. Points are the same as stock Q2 particles. As a demonstration
and/or testbed of the other two types, I tinkered up new effects for the
EF_TELEPORTER and EF_BLASTER effect flags.