Query resolutions from system for menu.
ref_gl now populates a cvar with a list of resolutions queried from the system, and the menu system uses that to populate the option in the video menu. (Bouncing off a cvar is kind of dopey but to do otherwise would break the DLL interface... not that we can't break the DLL interface, but I'd rather it stay vanilla for now.)

While preexisting settings for gl_mode will continue to be respected, making changes through the menu system will set gl_mode to -1 in favor of the vid_width/vid_height cvars.