Optimize the byte-swapping functions.
For all the focus on performance, it seems kind of weird to me that Quake2's byte-swapping functions bounced off a function pointer. I guess that's good for if your endianness were to somehow change at runtime?

Whatever. Use MSVC intrinsics if available. Otherwise fall back to the standard mask-and-shift approach.